+++ Packages/UGF.Code.Generate/Editor/Container/CodeGenerateContainerEditorUtility.Deprecated.cs
@@ -86,5 +86,72 @@ namespace UGF.Code.Generate.Editor.Container
 
             return container;
         }
+
+        /// <summary>
+        /// Gets public non static fields from the specified type.
+        /// </summary>
+        /// <param name="type">The type to get fields from.</param>
+        [Obsolete("GetFields has been deprecated.")]
+        public static FieldInfo[] GetFields(Type type)
+        {
+            if (type == null) throw new ArgumentNullException(nameof(type));
+
+            return type.GetFields(BindingFlags.Instance | BindingFlags.Public);
+        }
+
+        /// <summary>
+        /// Gets public non static properties from the specified type.
+        /// </summary>
+        /// <param name="type">The type to get properties from.</param>
+        [Obsolete("GetProperties has been deprecated.")]
+        public static PropertyInfo[] GetProperties(Type type)
+        {
+            if (type == null) throw new ArgumentNullException(nameof(type));
+
+            return type.GetProperties(BindingFlags.Instance | BindingFlags.Public);
+        }
+
+        /// <summary>
+        /// Determines whether the specified field is valid to generate field container.
+        /// <para>
+        /// The field can be valid for generating field container only if:
+        /// <para>- Field is public.</para>
+        /// <para>- Field is not static.</para>
+        /// <para>- Field is not readonly or const.</para>
+        /// </para>
+        /// </summary>
+        /// <param name="field">The field to validate.</param>
+        [Obsolete("IsValidField has been deprecated.")]
+        public static bool IsValidField(FieldInfo field)
+        {
+            if (field == null) throw new ArgumentNullException(nameof(field));
+
+            return field.IsPublic && !field.IsStatic && !field.IsLiteral && !field.IsInitOnly;
+        }
+
+        /// <summary>
+        /// Determines whether the specified property is valid to generate field container.
+        /// <para>
+        /// The property can be valid for generating field container only if:
+        /// <para>- Property is public.</para>
+        /// <para>- Property is not static.</para>
+        /// <para>- Property is not abstract.</para>
+        /// <para>- Property is can read and write.</para>
+        /// </para>
+        /// </summary>
+        /// <param name="property">The property to validate.</param>
+        [Obsolete("IsValidProperty has been deprecated.")]
+        public static bool IsValidProperty(PropertyInfo property)
+        {
+            if (property == null) throw new ArgumentNullException(nameof(property));
+
+            MethodInfo get = property.GetMethod;
+            MethodInfo set = property.SetMethod;
+
+            bool isGetValid = get != null && get.IsPublic && !get.IsStatic && !get.IsAbstract;
+            bool isSetValid = set != null && set.IsPublic && !set.IsStatic && !set.IsAbstract;
+
+            return property.GetIndexParameters().Length == 0 && isGetValid && isSetValid;
+        }
     }
 }